+++ src/tbkp_dp.c
@@ -5,7 +5,7 @@
 #include <time.h>
 
 static void tbkp_dp_timebombtable_set_on_active_row(
-    uint_fast32_t d, uint_fast32_t v, float val,
+    uint_fast32_t d, uint_fast32_t v, double val,
     TBKPDPTimeBombTable *const tb_t,
     const TBKPInstance *const inst
 ) {
@@ -19,7 +19,7 @@ static void tbkp_dp_timebombtable_set_on_active_row(
 }
 
 static void tbkp_dp_timebombtable_set_on_new_row(
-    uint_fast32_t d, uint_fast32_t v, float val,
+    uint_fast32_t d, uint_fast32_t v, double val,
     TBKPDPTimeBombTable *const tb_t,
     const TBKPInstance *const inst
 ) {
@@ -32,7 +32,7 @@ static void tbkp_dp_timebombtable_set_on_new_row(
     tb_t->rows[1u - tb_t->active_row][index] = val;
 }
 
-static float tbkp_dp_timebombtable_get(
+static double tbkp_dp_timebombtable_get(
     uint_fast32_t d, uint_fast32_t v, uint_fast32_t j,
     const TBKPDPTimeBombTable *const tb_t,
     const TBKPInstance *const inst
@@ -46,7 +46,7 @@ static float tbkp_dp_timebombtable_get(
     return tb_t->rows[row_idx][index];
 }
 
-static float tbkp_dp_timebombtable_getopt(
+static double tbkp_dp_timebombtable_getopt(
     uint_fast32_t d, uint_fast32_t v,
     const TBKPDPTimeBombTable *const tb_t,
     const TBKPInstance *const inst
@@ -68,18 +68,18 @@ static _Bool tbkp_dp_timebombtable_compute(
     for(uint_fast32_t j = 1u; j <= inst->n_tb_items; ++j) {
         const uint_fast32_t wj = inst->weights[j];
         const uint_fast32_t pj = inst->profits[j];
-        const float pij = inst->probabilities[j];
+        const double pij = inst->probabilities[j];
 
         for(uint_fast32_t dd = 0u; dd <= c; ++dd) {
             const uint_fast32_t d = c - dd;
 
             for(uint_fast32_t v = 0u; v <= U; ++v) {
-                const float current_val = tbkp_dp_timebombtable_get(d, v, j - 1u, tb_t, inst);
+                const double current_val = tbkp_dp_timebombtable_get(d, v, j - 1u, tb_t, inst);
 
                 if(wj > d || pj > v) {
                     tbkp_dp_timebombtable_set_on_new_row(d, v, current_val, tb_t, inst);
                 } else {
-                    const float new_val = tbkp_dp_timebombtable_get(d - wj, v - pj, j - 1u, tb_t, inst) * pij;
+                    const double new_val = tbkp_dp_timebombtable_get(d - wj, v - pj, j - 1u, tb_t, inst) * pij;
 
                     if(new_val > current_val) {
                         tbkp_dp_timebombtable_set_on_new_row(d, v, new_val, tb_t, inst);
@@ -246,7 +246,7 @@ void tbkp_dp_deterministictable_free(TBKPDPDeterministicTable* d_t) {
     free(d_t->t); d_t->t = NULL;
 }
 
-float tbkp_dp_solve(const TBKPInstance *const inst, TBKPDPStats* stats, const TBKPParams *const params) {
+double tbkp_dp_solve(const TBKPInstance *const inst, TBKPDPStats* stats, const TBKPParams *const params) {
     const uint_fast32_t U = inst->tb_ub_packed_profit;
     TBKPDPTimeBombTable tb_t = tbkp_dp_timebombtable_init(inst);
     TBKPDPDeterministicTable d_t = tbkp_dp_deterministictable_init(inst);
@@ -260,7 +260,7 @@ float tbkp_dp_solve(const TBKPInstance *const inst, TBKPDPStats* stats, const TB
     timeout = tbkp_dp_timebombtable_compute(&tb_t, start_time, inst, params);
 
     if(timeout) {
-        return -1.0f;
+        return -1.0;
     }
 
     stats->built_timebomb_table = true;
@@ -268,18 +268,18 @@ float tbkp_dp_solve(const TBKPInstance *const inst, TBKPDPStats* stats, const TB
     timeout = tbkp_dp_deterministictable_compute(&d_t, start_time, inst, params);
 
     if(timeout) {
-        return -1.0f;
+        return -1.0;
     }
 
     stats->built_deterministic_table = true;
 
-    float best_sol = -1.0f;
+    double best_sol = -1.0;
 
     for(uint_fast32_t d = 0; d <= inst->capacity; ++d) {
         const uint_fast32_t det_sol = tbkp_dp_deterministictable_getopt(inst->capacity - d, &d_t, inst);
 
         for(uint_fast32_t v = 0; v <= U; ++v) {
-            float sol = ((float)v + (float)det_sol);
+            double sol = ((double)v + (double)det_sol);
             sol *= tbkp_dp_timebombtable_getopt(d, v, &tb_t, inst);
 
             if(sol > best_sol) {
