+++ src/tbkp_dp.h
@@ -6,7 +6,7 @@
 #include "tbkp_instance.h"
 
 typedef struct {
-    float* rows[2];
+    double* rows[2];
     size_t active_row;
     uint_fast32_t last_index;
 } TBKPDPTimeBombTable;
@@ -21,6 +21,6 @@ void tbkp_dp_timebombtable_free(TBKPDPTimeBombTable* tb_t);
 TBKPDPDeterministicTable tbkp_dp_deterministictable_init(const TBKPInstance* inst);
 void tbkp_dp_deterministictable_free(TBKPDPDeterministicTable* d_t);
 
-float tbkp_dp_solve(const TBKPInstance *const inst, TBKPDPStats* stats, const TBKPParams *const params);
+double tbkp_dp_solve(const TBKPInstance *const inst, TBKPDPStats* stats, const TBKPParams *const params);
 
 #endif
